requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
fake-headers==1.0.2
fake-useragent==0.1.11
html5lib==1.1
idna==2.10
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23
urllib3==1.26.3
webencodings==0.5.1
